Fix incorrectly-imported OSS-Fuzz issues
OSV-2018-13, OSV-2024-417, OSV-2024-430, and OSV-2024-432 were all
OSS-Fuzz infrastructure regressions. When OSS-Fuzz breaks MSan, which is
unfortunately quite often, the result is false positive reports. OSV
seems to incorrectly classify these.

OSV-2018-206 was a bug in the fuzzer, not a bug in the library.

OSV-2023-41 was a bug in the library, but one we do not consider to be a
security bug as this code is not safe for use with untrusted inputs, as
documented in
https://commondatastorage.googleapis.com/chromium-boringssl-docs/x509.h.html#Deprecated-config-based-extension-creation

(The regression range is also wrong because it's flagging when the
fuzzer was added, but I've left that alone.)