Add standalone script for managing OSS-Fuzz -> OSV.
This adds the start of standalone infrastructure to manage syncing of
state between OSS-Fuzz issues and the OSS-Fuzz OSV feed.

This replaces existing functionality inside ClusterFuzz to send requests
to OSV, which had issues with reliability.

Instead, this script queries the OSS-Fuzz issue tracker as the source of
truth to generate requests to OSV.dev. In the future, we'll move the
bisection and OSS-Fuzz-Vulns record management functionality out of
OSV.dev.

This is part of https://github.com/google/osv.dev/issues/2892.